move method back onto controller
