+++ web/service/src/main/java/org/mobilitydata/gtfsvalidator/web/service/util/ValidationHandler.java
@@ -1,14 +1,10 @@
 package org.mobilitydata.gtfsvalidator.web.service.util;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.File;
-import java.util.Base64;
 import org.mobilitydata.gtfsvalidator.input.CountryCode;
 import org.mobilitydata.gtfsvalidator.runner.ValidationRunner;
 import org.mobilitydata.gtfsvalidator.runner.ValidationRunnerConfig;
 import org.mobilitydata.gtfsvalidator.util.VersionResolver;
-import org.mobilitydata.gtfsvalidator.web.service.controller.GoogleCloudPubsubMessage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -41,24 +37,5 @@ public class ValidationHandler {
     var config = configBuilder.build();
     runner.run(config);
     return outputPath;
-  }
-
-  /**
-   * Extracts the job ID and input file name from the Pub/Sub message.
-   *
-   * @param message
-   * @return
-   * @throws JsonProcessingException
-   */
-  public ValidationJobMetaData getFeedFileMetaData(GoogleCloudPubsubMessage.Message message)
-      throws JsonProcessingException {
-    var data = new String(Base64.getDecoder().decode(message.getData()));
-
-    var map = new ObjectMapper();
-    var node = map.readTree(data);
-
-    var inputFilename = node.get("name").textValue();
-    var jobId = inputFilename.split("/")[0];
-    return new ValidationJobMetaData(jobId, inputFilename);
   }
 }
